refactor: move LessonDetails and Less onOverviewPage
